refactor(init): Split fns out into own files
