+++ src/commands/init/remove-empty-config-values.js
@@ -1,0 +1,6 @@
+function removeEmptyConfigValues(configObj) {
+  Object.entries(configObj).forEach(([key, value]) => {
+    if (!value) delete configObj[key]
+  })
+}
+module.exports = removeEmptyConfigValues
